backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "latinos-backend"
version = "0.1.0"
description = "FastAPI backend for Latinos app"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic[email]",
    "sqlalchemy",
    "psycopg2-binary",
    "python-multipart",
    "python-jose[cryptography]",
    "python-dotenv",
    "passlib[bcrypt]",
    "httpx",
    "yfinance",
    "redis",
    "stripe",
    "cachetools",
    "argon2-cffi",
    "alpaca-trade-api",
    "slowapi"
]

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy"
]
